third_party/blink/web_tests/fast/frames/iframe-onload-and-domnodeinserted-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
This test ensures that any tree mutation in the load event handler cannot harm the tree consistency.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS loadEventFired is true
PASS unless crash.
PASS successfullyParsed is true

TEST COMPLETE
